+++ packages/bluesky/lib/src/bluesky.dart
@@ -21,7 +21,7 @@ sealed class Bluesky {
     final String? service,
     final String? relayService,
     final Duration? timeout,
-    final core.RetryConfig? retryConfig,
+    final core.RetryStrategy? retryConfig,
     final core.GetClient? getClient,
     final core.PostClient? postClient,
   }) => _Bluesky(
@@ -58,7 +58,7 @@ sealed class Bluesky {
     final String? service,
     final String? relayService,
     final Duration? timeout,
-    final core.RetryConfig? retryConfig,
+    final core.RetryStrategy? retryConfig,
     final core.GetClient? getClient,
     final core.PostClient? postClient,
   }) => _Bluesky(
@@ -98,7 +98,7 @@ sealed class Bluesky {
     final String? service,
     final String? relayService,
     final Duration? timeout,
-    final core.RetryConfig? retryConfig,
+    final core.RetryStrategy? retryConfig,
     final core.GetClient? getClient,
     final core.PostClient? postClient,
   }) => Bluesky.fromOAuth(
@@ -120,7 +120,7 @@ sealed class Bluesky {
     final String? service,
     final String? relayService,
     final Duration? timeout,
-    final core.RetryConfig? retryConfig,
+    final core.RetryStrategy? retryConfig,
     final core.GetClient? getClient,
     final core.PostClient? postClient,
   }) => _Bluesky(